chore(Flow): Move Flow related files to EmptyState
With the upcoming rework on the New Flow aspect of Kaoto, the existing
functionality will be only used in the Empty State, therefore, we're
moving those files to be a subcomponent of said component.

relates: https://github.com/KaotoIO/kaoto/issues/1030
